requirements.txt
pytest==6.2.5
Appium-Python-Client==3.1.1